+++ DXBase/DXBase/sources/Actor/Body/BoundingSegment.cpp
@@ -53,7 +53,7 @@ void BoundingSegment::draw(Matrix inv) const {
 	Vector3 pos0 = Vector3(component_.point[0].x, component_.point[0].y) * inv;
 	Vector3 pos1 = Vector3(component_.point[1].x, component_.point[1].y) * inv;
 
-	DrawLine(pos0.x, pos0.y, pos1.x, pos1.y, GetColor(255, 0, 0));
+	DrawLine(static_cast<int>(pos0.x), static_cast<int>(pos0.y), static_cast<int>(pos1.x), static_cast<int>(pos1.y), GetColor(255, 0, 0));
 
 	//DrawBox(component_.point[0].x, component_.point[0].y,
 	//	component_.point[3].x, component_.point[3].y, GetColor(255, 0, 0), FALSE);
@@ -75,9 +75,9 @@ void BoundingSegment::draw(int spriteID, Matrix inv) const {
 	}
 
 
-	DrawLine(pos0.x, pos0.y, pos1.x, pos1.y, GetColor(255, 0, 0));
+	DrawLine(static_cast<int>(pos0.x), static_cast<int>(pos0.y), static_cast<int>(pos1.x), static_cast<int>(pos1.y), GetColor(255, 0, 0));
 
-	DrawModiGraph(pos1.x, pos1.y, pos0.x, pos0.y, pos0.x, pos0.y, pos0.x, pos1.y,  spriteID, TRUE);
+	DrawModiGraph(static_cast<int>(pos1.x), static_cast<int>(pos1.y), static_cast<int>(pos0.x), static_cast<int>(pos0.y), static_cast<int>(pos0.x), static_cast<int>(pos0.y), static_cast<int>(pos0.x), static_cast<int>(pos1.y),  spriteID, TRUE);
 	//DrawBox(component_.point[0].x, component_.point[0].y,
 	//	component_.point[3].x, component_.point[3].y, GetColor(255, 0, 0), FALSE);
 	//DrawSphere3D(Vector3::Vector3ToVECTOR(component_.center_), component_.radius_, 32, GetColor( 255,0,0 ), GetColor( 255, 255, 255 ), TRUE ) ;
